fix!: Replace chrono with time (#460)
This PR replaces the unmaintained `chrono` crate with `time` crate. One missing feature in `time` compared to `chrono` is that `time` has its own [format description](https://time-rs.github.io/book/api/format-description.html) that is incompatible with `chrono`, and thus liquid templates, so this PR adds a super simple formatter that replicates the `chrono` behavior. In true Rust fashion once it compiled it worked and all of the tests passed, but I will add some additional tests tomorrow for some niche cases that I noticed were missing from the current tests. I also noticed that chrono was lying in a few cases about what it supported (eg. `%Z` to show the time zone name is not actually implemented in chrono) so in those cases I preserved the `chrono` behavior even when it doesn't match up to the format specifiers [supposedly](https://shopify.github.io/liquid/filters/date/) supported by liquid.

Resolves: #446

+++ tests/test_helper.rs
@@ -1,7 +1,7 @@
 pub use liquid_core::Value::Nil;
 
 #[allow(dead_code)]
-pub fn date(y: i32, m: u32, d: u32) -> liquid_core::Value {
+pub fn date(y: i32, m: u8, d: u8) -> liquid_core::Value {
     use liquid_core::model::Date;
     use liquid_core::model::Value;
     Value::scalar(Date::from_ymd(y, m, d))
